Use setup.sh to install polymer-build#df7b17c for gulp build
This must be done in a script instead of via package.json because
npm seems to install the official release instead of the requested
commit from package.json. This works fine when the install command
is run explicitly.

The setup script uses npm to pre-install polymer-build#df7b17c
(downloads the source). Then it builds it, and links it to
polymer-cli.